# Minor change to allow the Tokenizer to accept a io.ByteReader