This makes it possible to use the Tokenizer with other types of Readers, not just strings.