@sougou 2017-09-18
TL;DR: This change can be accepted now, but there is a big revamp coming that will break this. Let us know what you think:

There is a proposal to rework the tokenizer and other related parts to improve performance. The background for it is #2849 filed by @demmer. The prototype is here: https://github.com/youtube/vitess/compare/master...sougou:perf_hack?expand=1. Can you look at the change to token.go? It basically gets rid of the reader altogether. Will this break you permanently, or will you be able to rework things to absorb the changes?