@bramp 2017-09-20
Oh ya, that would totally break me. The NewStringTokenizer now stores the SQL string, and just uses indexes with it.

In my case I don't know how long the SQL string is since I'm reading dozens of statements from a file. Currently I can use the parser with a io.Reader, and some minor hackier to read one Statement at a time.

From scanning your branch it wouldn't be too difficult to put the io.Reader back. I don't see why it would cause any more allocations than you already do.