@sougou 2017-09-22
This is really strange. I did comparison benchmarks. Here are the results:

vtgate:
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkWithNormalizer-4              5507854       9407342       +70.80%
BenchmarkWithoutNormalizer-4           9849          9795          -0.55%
BenchmarkResolveKeyRangeToShards-4     139           142           +2.16%
```

But vttablet hasn't changed much:
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkExecuteVarBinary-4      5090057       4914972       -3.44%
BenchmarkExecuteExpression-4     1213767       1140250       -6.06%
```

I think that an abstraction layer at the level of a byte operation has weird overheads. I'm instead thinking that we should carve out something at a higher level. That way, we can take care of this issue as well as the continuous streaming requirement that you have from #3191.