@sougou 2017-10-11
Since the overhead introduced is too high, I'll be closing this issue.