@bramp 2017-10-12
Just some other minor comments. I noticed the slow down seems to be because when using a `*strings.Reader` the ReadByte() and Tokenizer.next() methods are inlined. When I switch it out with a bufio.Reader, neither are inlined, and that seems to cause (as you noted) a factor of two performance regression. In particular time is spent in scanString() when parsing really long strings (such as in https://github.com/youtube/vitess/blob/master/go/vt/vtgate/bench_test.go#L40).  Feel free to close this PR, but I'm just working on a rewrite for scanString, that uses bufio.ReadBytes() and improves performance in both the strings.Reader and io.Reader case. However just fixing some edge cases to make it correct.